Creating the universal executor
